Split UI files
